refactor(executor): reorganize tests into focused test modules
- Created src/executor/tests/ directory with 5 focused test modules:
  * expansion_tests.rs - tests for variable expansion and wildcards
  * redirection_tests.rs - tests for FD operations and redirections
  * command_tests.rs - tests for command execution and pipelines
  * subshell_tests.rs - placeholder for subshell-specific tests
  * execution_tests.rs - tests for main execute() function and control flow

- Reduced executor/mod.rs from 2,844 lines to 679 lines (core logic only)
- All 555+ tests pass successfully
- Improved test organization and maintainability

This completes Phase 1 of the executor refactoring, achieving the goal
of separating tests from implementation for better code organization.
