+++ src/executor/tests/mod.rs
@@ -1,0 +1,9 @@
+//! Test modules for the executor
+//!
+//! This module organizes tests into focused areas matching the executor's module structure.
+
+mod expansion_tests;
+mod redirection_tests;
+mod command_tests;
+mod subshell_tests;
+mod execution_tests;
