+++ src/executor/tests/subshell_tests.rs
@@ -1,0 +1,13 @@
+//! Tests for subshell execution and compound commands
+
+use crate::state::ShellState;
+
+// Note: Subshell-specific tests will be added here as the subshell module evolves.
+// Currently, subshell functionality is tested indirectly through execution_tests.rs
+
+#[test]
+fn test_subshell_placeholder() {
+    // Placeholder test to ensure module compiles
+    let _shell_state = ShellState::new();
+    assert!(true);
+}
